+++ src/constants/index.js
@@ -342,3 +342,46 @@ export const BOOTSTRAP_DEFAULTS = {
   COLLAPSE_TRANSITION_TIME: 300,
   MODAL_ZINDEX: 1055,
 };
+
+export const TIME_OF_DAY_PERIODS = [
+  {
+    rangeString: '12:01 - 15:00',
+    rangeMinutesMin: 721,
+    rangeMinutesMax: 900,
+  },
+  {
+    rangeString: '15:01 - 18:00',
+    rangeMinutesMin: 901,
+    rangeMinutesMax: 1080,
+  },
+  {
+    rangeString: '18:01 - 21:00',
+    rangeMinutesMin: 1081,
+    rangeMinutesMax: 1260,
+  },
+  {
+    rangeString: '21:01 - 00:00',
+    rangeMinutesMin: 1261,
+    rangeMinutesMax: 1440,
+  },
+  {
+    rangeString: '00:01 - 03:00',
+    rangeMinutesMin: 1,
+    rangeMinutesMax: 180,
+  },
+  {
+    rangeString: '03:01 - 06:00',
+    rangeMinutesMin: 181,
+    rangeMinutesMax: 360,
+  },
+  {
+    rangeString: '06:01 - 09:00',
+    rangeMinutesMin: 361,
+    rangeMinutesMax: 54,
+  },
+  {
+    rangeString: '09:01 - 12:00',
+    rangeMinutesMin: 55,
+    rangeMinutesMax: 720,
+  }
+];
